+++ sdk/include/opentelemetry/sdk/trace/tracer_config.h
@@ -16,7 +16,7 @@ namespace trace
  * ScopeConfigurator should be used to compute the desired TracerConfig which can then be used to
  * configure a Tracer.
  */
-class TracerConfig
+class OPENTELEMETRY_EXPORT TracerConfig
 {
 public:
   bool operator==(const TracerConfig &other) const noexcept;
@@ -51,8 +51,6 @@ public:
 private:
   explicit TracerConfig(const bool disabled = false) : disabled_(disabled) {}
   bool disabled_;
-  static const TracerConfig kDefaultConfig;
-  static const TracerConfig kDisabledConfig;
 };
 }  // namespace trace
 }  // namespace sdk