+++ sdk/src/logs/logger_context.cc
@@ -6,6 +6,8 @@
 #include <utility>
 #include <vector>
 
+#include "opentelemetry/sdk/instrumentationscope/scope_configurator.h"
+#include "opentelemetry/sdk/logs/logger_config.h"
 #include "opentelemetry/sdk/logs/logger_context.h"
 #include "opentelemetry/sdk/logs/multi_log_record_processor.h"
 #include "opentelemetry/sdk/logs/processor.h"
@@ -19,10 +21,13 @@ namespace logs
 {
 
 LoggerContext::LoggerContext(std::vector<std::unique_ptr<LogRecordProcessor>> &&processors,
-                             const opentelemetry::sdk::resource::Resource &resource) noexcept
+                             const opentelemetry::sdk::resource::Resource &resource,
+                             std::unique_ptr<instrumentationscope::ScopeConfigurator<LoggerConfig>>
+                                 logger_configurator) noexcept
     : resource_(resource),
       processor_(
-          std::unique_ptr<LogRecordProcessor>(new MultiLogRecordProcessor(std::move(processors))))
+          std::unique_ptr<LogRecordProcessor>(new MultiLogRecordProcessor(std::move(processors)))),
+      logger_configurator_(std::move(logger_configurator))
 {}
 
 void LoggerContext::AddProcessor(std::unique_ptr<LogRecordProcessor> processor) noexcept
@@ -39,6 +44,12 @@ LogRecordProcessor &LoggerContext::GetProcessor() const noexcept
 const opentelemetry::sdk::resource::Resource &LoggerContext::GetResource() const noexcept
 {
   return resource_;
+}
+
+const instrumentationscope::ScopeConfigurator<LoggerConfig> &LoggerContext::GetLoggerConfigurator()
+    const noexcept
+{
+  return *logger_configurator_;
 }
 
 bool LoggerContext::ForceFlush(std::chrono::microseconds timeout) noexcept